Fixed LoginTest() + Refactored Menu + Bank() changes
LoginTest now renamed to Login, it returns a boolean so we can determine if an account exists and if it was successful or not.
Bank() method login returns boolean also to be used in BankApp() to see if login was successful, it then uses the outcome to move onto the next menu
Menu also changed to be smoother and more user friendly, no longer immediately exits in second loop, asks user if they want to logout instead and brings them to main menu.